refactor : DataSource 적용 및 repository 통합

+++ app/src/main/java/woowacourse/shopping/data/product/JsonMapper.kt
@@ -19,7 +19,7 @@ fun Product.toJson(): String {
 
 fun JSONObject.toProduct(): Product {
     return Product(
-        id = getInt("id"),
+        id = getLong("id"),
         name = getString("name"),
         imageUrl = getString("imageUrl"),
         price = Price(getInt("price")),
